+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Added
 - add long options for commands to improve readability
 - add descriptions to commands
+- parameter for port
+- parameter to disable use of TLS encryption
+
+### Changed
+- change default call to check via TLS on port 993
 
 ## [0.0.2] - 2015-07-14
 ### Changed